Working 'Hello World' kernel in D
This commit adds a core for a hello world test.
It uses Adam D. Ruppes minimal d runtime which is located at
http://arsdnet.net/dcode/minimal.zip

It uses my build system called 'Wild', located at
https://github.com/Vild/Wild
It expects the wild binary to be located in the root directory
of PowerNex, then you just need to run `./build` to build.

Signed-off-by: Dan Printzell <xwildn00bx@gmail.com>