Big commit, Sorry
This adds the concept of Consoles and Framebuffers.
You can write UTF8 text to a console and it will display it via a
a VirtualConsoleScreen or a the serial port. If you read from it
you will get input from the keyboard or the serial port.

VirtualConsole are emulated consoles. This means that you can have more
VirtualConsole than you have physical consoles.
Physical consoles include TextMode, BGA and Serial.

A VirtualConsoleScreen is the part that the VirtualConsole write to, to
display text on the screen. It can either be via a Framebuffer or via
the TextMode.

A Framebuffer is a array pixels with a height and width.

The ConsoleManager control the VirtualConsoles. For example, gives them
keyboard input and changes which one that should be active and write to
its framebuffer.

There are also a lot of VFS changes and bug fixes that where required.

Now the ConsoleFont is compiled into the kernel.

Signed-off-by: Dan Printzell <xwildn00bx@gmail.com>